+++ forge20/src/main/java/com/vecoo/extralib/gson/UtilGson.java
@@ -73,7 +73,7 @@ public abstract class UtilGson {
             writer.close();
             return true;
         } catch (Exception e) {
-            ExtraLib.getLogger().error("Write file sync error");
+            ExtraLib.getLogger().error("Write file sync error.");
             return false;
         }
     }
@@ -128,7 +128,7 @@ public abstract class UtilGson {
             callback.accept(data.toString());
             return true;
         } catch (Exception e) {
-            ExtraLib.getLogger().error("Read file sync error");
+            ExtraLib.getLogger().error("Read file sync error.");
             return false;
         }
     }