+++ changelog.md
@@ -1,1 +1,5 @@
-* Fix load mod.
+* Update gradle depends.
+* Added server and command utils.
+* Added fast execute command.
+* Added save and load ItemStack.
+* Improve logger.